disable concurrency
While the documentation states that pending jobs for previous workflows are
cancelled when concurrency is used, when deployments are used jobs requiring
approval are not cancelled until they are approved (design defect?). Since
terraform already has a mechanism to prevent concurrent apply invocations, we
can remove the use of github action concurrency.